+++ services/celest_cloud_auth/CHANGELOG.md
@@ -3,6 +3,7 @@
 - feat!: Make database modular
 - refactor!: Rename `AuthDatabase` to `CloudAuthDatabase`
 - refactor!: Rename all entities with `cloud_auth_` prefix
+- fix: Correctly route `users/me`
 - chore: Make `Authorizer` a class
 - chore: Update dependencies